Pagination Fix
1.

The 'Accept: application/vnd.github.preview' header must be passed for
the Search API to work during the preview period.

'next_ops' is created to hold a copy of the headers to be used in
subsequent get() requests, as 'opts' is cleared by the initial
request() call.

This permits pagination during the Search API's preview period.

2.

request() returns a Sawyer, so the Array we are interested in is
data.items.